주문 및 상품 관련 AOP 로깅 기능 추가 및 유효성 검사 구현
- AOP 로깅 클래스(LoggingAspect) 추가하여 Controller에 로깅 기능 구현
- OrderRequestDto에 주문 생성 시 유효성 검사 로직 추가
- 주문 생성 및 업데이트 관련 예외 처리 로직 추가
- 상품(Product) 및 주문(Orders) 관련 엔티티와 레포지토리 구조 업데이트

+++ shop/src/main/java/com/shop/shop/domain/order/OrderService.java
@@ -9,8 +9,8 @@ import java.util.List;
 public interface OrderService {
 
     OrderResponseDto createOrder(OrderRequestDto requestDto);
-    void cancelOrder(Integer orderId);
-    OrderResponseDto getOrder(Integer orderId);
+    void cancelOrder(Long orderId);
+    OrderResponseDto getOrder(Long orderId);
     List<OrderResponseDto> getAllOrders();
-    OrderResponseDto updateOrder(Integer orderId, OrderRequestDto requestDto);
+    OrderResponseDto updateOrder(Long orderId, OrderRequestDto requestDto);
 }
